feature/#3 : properties를 통한 vlm 설정 관리 / dto 추가를 고려한 entity -> domain으로 디렉토리 변경(TDD - Refactor Stage)
- model.properties를 통한 vlm 설정 관리 -> 추후 env를 통해 운영 시 변경
  가능하도록 구성
- request/response 클래스 추가를 고려한 entity -> domain으로 디렉토리명
  변경
