+++ src/main/kotlin/zzibu/jeho/tagify/config/ModelConfig.kt
@@ -1,0 +1,18 @@
+package zzibu.jeho.tagify.config
+
+import org.springframework.boot.context.properties.EnableConfigurationProperties
+import org.springframework.boot.context.properties.bind.ConstructorBinding
+import org.springframework.context.annotation.Bean
+import org.springframework.context.annotation.Configuration
+import org.springframework.context.annotation.PropertySource
+
+@Configuration
+@PropertySource("classpath:model.properties")
+@EnableConfigurationProperties(ModelProperties::class)
+class ModelConfig(val modelProperties: ModelProperties) {
+
+    @Bean
+    fun assistantMessage() : String{
+        return modelProperties.message
+    }
+}
