+++ src/main/kotlin/zzibu/jeho/tagify/service/TagService.kt
@@ -10,19 +10,15 @@ import org.springframework.ai.chat.model.ChatModel
 import org.springframework.ai.chat.model.ChatResponse
 import org.springframework.ai.chat.prompt.Prompt
 import org.springframework.ai.ollama.api.OllamaOptions
-import org.springframework.core.io.ClassPathResource
 import org.springframework.core.io.InputStreamResource
 import org.springframework.core.io.Resource
 import org.springframework.stereotype.Service
 import org.springframework.util.MimeTypeUtils
 import org.springframework.web.multipart.MultipartFile
 import zzibu.jeho.tagify.domain.TagInfo
-import zzibu.jeho.tagify.domain.dto.GenRequest
 import zzibu.jeho.tagify.repository.TagRepository
 import java.io.IOException
-import java.net.URL
 import java.util.List
-
 
 private val logger = KotlinLogging.logger{}
 
@@ -30,16 +26,14 @@ private val logger = KotlinLogging.logger{}
 class TagService(
     private val tagRepository: TagRepository,
     private val chatModel: ChatModel,
+    private val assistantMessage: String,
     ) {
-
-    private val assistantMessage : String =  """
-                Look at the image and list the words that come to mind in an array 
-                    format : [1, 2, 3, 4, 5]
-            """.trimIndent()
-
     fun generateTagByImage(image:MultipartFile,  name : String,url : String, owner : String): TagInfo {
         val vlmResponse = sendImageToVLM(image)
         val tags = jsonToList(vlmResponse)
+
+        logger.info { tags }
+
         val tagInfo = TagInfo(
             id = null,
             url = url,
@@ -56,15 +50,16 @@ class TagService(
 
         val userMessage = UserMessage(
             assistantMessage,
-            List.of<Media>(Media(MimeTypeUtils.IMAGE_PNG, imageData))
+            List.of<Media>(Media(MimeTypeUtils.ALL, imageData))
         )
-        logger.info { assistantMessage }
+
         val response: ChatResponse = chatModel.call(
             Prompt(List.of<Message>(userMessage), OllamaOptions.create().withModel("llava"))
         )
-        return response.result.output.content
+        return response.result.output.content.trimIndent()
     }
 
+    // 모델 반환 값을 json으로 설정하였으므로, 이를 리스트 형태로 변경
     private fun jsonToList(jsonString : String) : kotlin.collections.List<String> {
         val objectMapper = jacksonObjectMapper()
         val map: Map<String, String> = objectMapper.readValue(jsonString)
@@ -75,6 +70,7 @@ class TagService(
         return tagList.toList()
     }
 
+    // Spring AI 프레임워크의 Media 생성자 타입 Resource로 맞추기 위함.
     @Throws(IOException::class)
     private fun convertToInputStreamResource(file: MultipartFile): Resource {
         return object : InputStreamResource(file.inputStream) {
