+++ src/main/resources/model.properties
@@ -1,0 +1,2 @@
+vlm.message=Look at the image and list the words that come to mind in an array \
+format : [1 : "tag1", 2: "tag2", 3: "tag3", 4: "tag4", 5: "tag5"]
